COMMIT OF THE DAY,
Fix: I changed the hasAnyAuthority to hasAnyRole in the security config.
Add: Since I added this, the security assumes that it should only grant authorization if someone logs in with the authority 'ROLE_...'. Therefore, I modified the existing role in my data.sql to 'ROLE_ADMIN' and 'ROLE_MECHANIC'. The 'ROLE' prefix indicates to the security which areas someone with that role is allowed or not allowed to access.
Delete: After modifying the security config, I also found an old method that I deleted.