Commit after a failed Postman test run.
Add: In the security config, I made some changes to the filter chain. I have decided to write out the full URL for each role.
Fix: Additionally, I encountered errors while copying the security. The issue was with the CustomerPasswordEncoder, which I removed from the constructor and set to autowired.
Fix: After that, I reviewed various classes and fixed errors. For example, in the CustomerUserDetailService, which was originally named UserDetailService, I found a mistake. I had mistakenly injected a UserRepository instead of a UserService. I fixed everything, and now I have access to my endpoint.
Fix: To my surprise, I noticed confusing method names in the SparkplugController. I corrected them, so now it's clear what the purpose of each method is.