Fixing problems and adding new things.
Fix: I fixed an issue with the JWTRequestFilter. I was using dependency injection of two services, a customUserDetailService and a jwtService. The problem was with the customUserDetailService, which was incorrectly referencing an existing UserDetailService interface. However, it has been fixed now.
Add: I further tested the endpoints of the car inspection classes, and I encountered a few things that didn't work smoothly.
Fix: I made various adjustments and continued testing. It's still not optimal, but I should make significant progress tomorrow.
Add: I added new fields to the car inspection input and output DTOs.
Fix: The name of the CustomUserDetailService was misspelled as CustomerUserDetailService.