Today I continued testing the endpoint.
Fix: I had issues creating a new object. After a lot of searching, I realized that I hadn't properly defined the variables in the body of Postman. I was continuously sending a body as a text file instead of JSON.
Fix: I discovered that the car inspection was missing a few logic components, which I added, and now it's working.
Fix: After the car inspection, I moved on to the car repair, but I hadn't yet added the 'add part to car repair' functionality. I worked hard on it, and it was quite a task to add it because the part class is polymorphic. Initially, I thought it would be convenient to create a method for adding each component to the repair class. Unfortunately, that didn't work because it expects a part type. So, here's what I did: I created a switch statement where, upon receiving a part controller, it scans the PartType enum, and then it redirects to the corresponding repository.
Fix: I changed 'SPARKPLUG' to 'SPARK_PLUG' in the PartType enum.