The addPartToCarRepair created.
Fix: Today, I looked into the functioning of the addPartToCarRepair, but after many attempts, I was unable to link a part to a car repair. After extensive research, I discovered that I needed to pass a part type and an ID number for the linkage.
Add: To address this, I created a DTO (Data Transfer Object) that needed to be used as input. After adding this component, I was able to send a body where I could make the linkage. The body included a partID and an enum that contained the partType.
Fix: Upon setting up the code and completing it, a syntax error arose. It indicated that it couldn't define the partType enum. After a lot of searching, I found that there was an error in the data.sql file. I had not written the part types in uppercase. After correcting that, the linkage worked correctly.