Today, I completed two service tests.
add: I tested the 'add' methods in the CarInspectionService and carService tests, which were missing in the entire system. I aimed for 100% line coverage for the car service, and I achieved it. For the carInspectionService, I reached 97% coverage. I know where the problem lies, and I'll think of a solution to handle the remaining 3%.
fix: Additionally, I modified the existing test methods to match the adjusted service methods.
fix: I started fixing an integration test, which is challenging to get back on track. However, I should be able to resolve it by tomorrow.
add: I added a new URL and tested it in Postman.