Today, I managed to get my integration tests working.
add: After encountering a few error messages and trying different approaches, someone suggested why I wasn't using an 'application-test.properties' file, as that seemed to be the source of my problem. Although I was receiving data, it wasn't the data I had created in my test method. Instead, it was the data stored in my main database and, furthermore, the data I had pre-defined in my 'data.sql' file. After creating the 'application-test.properties' file, everything started working properly again.

add: I added the dependency for the H2 database in my 'pom.xml' file. H2 is a lightweight, in-memory database. The advantage of using it is that tests run quite fast, and it keeps the test data separate from the rest. The dependency is located under the groupId 'com.h2database.'